algo2

+++ ProjectActivFormationParis/src/com/formation/etga/exo/ExoAlgo2.java
@@ -20,7 +20,7 @@ public class ExoAlgo2 implements ImpAlgo2 {
 		int nbDiviseur = 0;
 		int racine = (int) Math.sqrt(n);
 
-		if ((n > 0) && ((float) n <= Integer.MAX_VALUE)) {
+		if ((n > 0) && ((long) n <= Integer.MAX_VALUE)) {
 			for (int i = 1; i <= racine; i++) {
 				if (n % i == 0) {
 					nbDiviseur += 2;
